Organize sink files
