+++ src/connector/src/sink/mod.rs
@@ -1,0 +1,31 @@
+// Copyright 2022 Singularity Data
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+pub mod mysql;
+
+use async_trait::async_trait;
+use risingwave_common::array::StreamChunk;
+use risingwave_common::catalog::Schema;
+use risingwave_common::error::Result;
+
+#[async_trait]
+pub trait Sink {
+    async fn write_batch(&mut self, chunk: StreamChunk, schema: &Schema) -> Result<()>;
+
+    fn endpoint(&self) -> String;
+    fn table(&self) -> String;
+    fn database(&self) -> Option<String>;
+    fn user(&self) -> Option<String>;
+    fn password(&self) -> Option<String>; // TODO(nanderstabel): auth?
+}
